refactor(aiprovider): Unified icon configuration to SmartHopperSettings
